Rename C++ Baseline.x filenames
The poorly named Baseline.cc and Baseline.h files have a wrapper baseline.cc that
is not related to the baseline.py module. While none of these names are ideal,
the sole class in Baseline.cc is BaselineUtils, so we rename the C++ files
BaselineUtils.x to avoid confusion.
