+++ include/lsst/meas/deblender/BaselineUtils.h
