+++ python/lsst/meas/deblender/SConscript
@@ -1,3 +1,3 @@
 # -*- python -*-
 from lsst.sconsUtils import scripts
-scripts.BasicSConscript.pybind11(['baseline'])
+scripts.BasicSConscript.pybind11(['baselineUtils'])
